Fix spacing and module naming
Module declaration was fixed.

Getting the WM_NAME function was commented out. I don't think it will
remain in this file, because it could be useful for testing purpose.
I will likely delete it and move it to its own file in a testing dir.

In other, more fun news, setWMName's let declaration had been throwing
some sort of parse error. It seemed to have to do with the spacing of
the do-block. Due some naming issues, files needed to be moved about
with git mv.